pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.116.1",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "scalar-fastapi>=1.2.2",
    "uvicorn>=0.35.0",
    "python-dotenv>=1.0.1",
    "sseclient-py>=1.8.0",
    "lark>=1.2.2",
    "opik>=1.4.5",
    "autoflake>=2.3.1",
    "flake8>=7.1.1",
    "black>=25.1.0",
    "mypy>=1.14.1",
    "isort>=6.0.0",
    "colorlog>=6.9.0",
    "fastapi[standard]>=0.115.6",
    "pydantic-settings>=2.7.0",
    "pre-commit>=4.0.1",
    "ruff>=0.14.10",
]
